requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
asgiref==3.8.1
asttokens==2.4.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.8.2
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.36.22
botocore==1.36.22
Brlapi==0.8.5
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
crispy-bootstrap5==2024.10
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==5.1.3
django-crispy-forms==2.3
dnf==4.22.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
Flask==3.0.3
fonttools==4.55.3
fros==1.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
humanize==4.9.0
idna==3.10
ipython==8.29.0
iso639==0.1.4
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
jsonpickle==3.4.2
kiwisolver==1.4.8
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
networkx==3.4.2
nftables==0.1
numpy==2.2.1
oauthlib==3.2.2
olefile==0.47
packaging==24.2
pandas==2.2.3
parso==0.8.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.2.2
ply==3.11
productmd==1.43
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
PuLP==2.9.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
pytz==2025.1
pyudev==0.24.3
pyvis==0.3.2
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpm==4.20.0
rsa==4.9
s3transfer==0.11.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.1
systemd-python==235
Tempita==0.5.2
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.21.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.2
xkbregistry==0.3
